requirements.txt
nltk==3.7
dm-haiku==0.0.9
jax==0.3.25
jaxlib @ https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.3.25+cuda11.cudnn82-cp37-cp37m-manylinux2014_x86_64.whl
einshape==1.0.0
dm-tree==0.1.7
parameterized==0.8.1
nose==1.3.7
ml-collections==0.1.1
optax==0.1.4
chex==0.1.5
more-itertools==9.0.0
tensorflow==2.9.1
tensorflow-datasets==4.6.0
